CRYPTOGRAPHY
Cryptography - Changing the plane message into something anyone cant understands except the person who is having a key to crack it. Or it is the art of writing or solving codes.
Cipher Text - A text which is the result of Encryption
Encryption - Converting a message to a secret message 
Decryption - Converting back the secret message to a normal message
Problem statement:-  Write a python program that converts, any given text to ciphertext, and when the key is available it should convert the ciphertext back to the normal message.